changed long to BigInteger, because JLibGDB is using BigInteger as default data type

+++ src/main/java/com/gkd/Global.java
@@ -1,9 +1,11 @@
 package com.gkd;
 
+import java.math.BigInteger;
+
 public class Global {
 	public static String version = "20130520";
 	public static boolean debug = true;
-	public static long osDebug = -1;
+	public static BigInteger osDebug = BigInteger.valueOf(-1);
 	public static String jmpLog = "jmp.log";
 	public static String osLog = "os.log";
 	public static String interruptLog = "interrupt.log";
